Let's use one condition instead of one billion
